Add collection variables page
