+++ src/common/components/Collections/CollectionVariables.tsx
@@ -1,140 +1,27 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-import { UseMutateFunction } from '@tanstack/react-query';
-import { PlusIcon, Trash2 } from 'lucide-react';
-import { ChangeEvent, useEffect, useMemo, useState } from 'react';
-import {
-	useForm,
-	useFieldArray,
-	Control,
-	UseFieldArrayRemove,
-	Controller,
-} from 'react-hook-form';
+import { PlusIcon } from 'lucide-react';
+import { useEffect, useMemo } from 'react';
+import { useForm, useFieldArray } from 'react-hook-form';
 import { useParams } from 'react-router-dom';
 
-// import { useDebounce } from 'use-debounce';
-
+import EnvironmentItem from '../Environments/EnvironmentItem';
 import { Button } from '../ui/button';
-import { Input } from '../ui/input';
 
+import { useCollectionsQuery } from '@/common/api/collections';
 import {
 	useCreateEnvironmentMutation,
 	useDeleteEnvironmentMutation,
-	useEditEnvironmentMutation,
 	useEnvironmentsQuery,
 } from '@/common/api/enviroments';
 import { Environment } from '@/common/types/environment';
 
-interface IEnviromentItem {
-	index: number;
-	enviroment: Environment;
-	collectionId: string;
-	control: Control<
-		{
-			environments: Environment[];
-		},
-		any
-	>;
-	removeItem: UseFieldArrayRemove;
-	deleteMutation: UseMutateFunction<
-		any,
-		Error,
-		string,
-		{
-			oldEnvs: Environment[] | undefined;
-		}
-	>;
-}
-
-function EnvironmentItem({
-	enviroment,
-	index,
-	collectionId,
-	control,
-	removeItem,
-	deleteMutation,
-}: IEnviromentItem) {
-	const [name, setName] = useState<string>(enviroment.name);
-	const [value, setValue] = useState<string>(enviroment.value);
-
-	// const [debouncedName] = useDebounce(name, 1000);
-	// const [debouncedValue] = useDebounce(value, 1000);
-
-	const { mutate: editEnviromentMutation } = useEditEnvironmentMutation({
-		collectionId,
-	});
-
-	const generateMutationData = (field: string, value: string) => {
-		const mutationData =
-			field === 'name'
-				? {
-						id: enviroment.id,
-						name: value,
-						value: enviroment.value,
-						collectionId,
-				  }
-				: { id: enviroment.id, name: enviroment.name, value, collectionId };
-
-		editEnviromentMutation(mutationData);
-	};
-
-	return (
-		<li
-			key={index}
-			className="flex flex-row items-center gap-2 pb-2"
-			data-test="collection-variables-input-list"
-		>
-			<Controller
-				render={({ field: nameField }) => (
-					<Input
-						placeholder="Name"
-						className="w-1/3"
-						data-test="collection-variables-input-name"
-						{...nameField}
-						onChange={(e: ChangeEvent<HTMLInputElement>) => {
-							nameField.onChange(e.target.value);
-							setName(e.target.value);
-							generateMutationData(nameField.name, name);
-						}}
-					/>
-				)}
-				name={`environments.${index}.name`}
-				control={control}
-			/>
-			<Controller
-				render={({ field: valueField }) => (
-					<Input
-						placeholder="Value"
-						className="w-2/3"
-						data-test="collection-variables-input-value"
-						{...valueField}
-						onChange={(e: ChangeEvent<HTMLInputElement>) => {
-							valueField.onChange(e.target.value);
-							setValue(e.target.value);
-							generateMutationData(valueField.name, value);
-						}}
-					/>
-				)}
-				name={`environments.${index}.value`}
-				control={control}
-			/>
-			<button
-				type="button"
-				className="font-semibold"
-				onClick={() => {
-					enviroment.id ? deleteMutation(enviroment.id) : removeItem(index);
-				}}
-			>
-				<Trash2 className="w-6 h-6 cursor-pointer text-slate-500 hover:text-primary" />
-			</button>
-		</li>
-	);
-}
-
 export default function CollectionVariables() {
 	const params = useParams();
+	const { data: collections } = useCollectionsQuery();
+
 	const collectionId = useMemo(() => {
 		return params.collectionId ? params.collectionId : '';
 	}, [params]);
+	const collection = collections?.find((col) => col.id === collectionId);
 
 	const { data: environmentsData, isLoading } = useEnvironmentsQuery({
 		collectionId,
@@ -155,7 +42,12 @@ export default function CollectionVariables() {
 	const variablesList = useMemo(() => {
 		return environmentsData ? environmentsData : [];
 	}, [environmentsData]);
-	console.log('variablesList', '\n', variablesList);
+
+	const { fields, append, remove } = useFieldArray({
+		keyName: 'key',
+		control,
+		name: 'environments',
+	});
 
 	useEffect(() => {
 		if (!isLoading) {
@@ -165,12 +57,6 @@ export default function CollectionVariables() {
 		}
 	}, [environmentsData, isLoading, reset, variablesList]);
 
-	const { fields, append, remove } = useFieldArray({
-		keyName: 'key',
-		control,
-		name: 'environments',
-	});
-	console.log('fields', '\n', fields);
 	const addNewInputVariable = () => {
 		append({
 			id: '',
@@ -185,9 +71,6 @@ export default function CollectionVariables() {
 		if (environments) {
 			environments.forEach((environment) => {
 				if (!environment.id && environment.name && environment.value) {
-					// console.log(`createEnviromentMutation(name: ${environment.name},
-					// 	value: ${environment.value},
-					// 	collectionId: ${collectionId})`);
 					createEnviromentMutation({
 						name: environment.name,
 						value: environment.value,
@@ -199,6 +82,7 @@ export default function CollectionVariables() {
 	};
 
 	if (isLoading) {
+		// eslint-disable-next-line @typescript-eslint/no-explicit-any
 		return 'Loading...' as any;
 	}
 
@@ -216,8 +100,11 @@ export default function CollectionVariables() {
 					>
 						Collections variables
 					</h1>
-					<p className="px-2 text-sm font-semibold text-slate-500">
-						Collection Name
+					<p
+						className="px-2 text-sm font-semibold text-slate-500"
+						data-test="collection-variables-collection-name"
+					>
+						{collection?.name}
 					</p>
 				</div>
 				<div>
